Hoist FilterNetexApp into lib to remove duplication
FilterNetexApp existed as two byte-identical copies (cli.app and
pipeline.app) that could silently drift. All of its dependencies already
live in netex-tools-lib, so move the single canonical class to
org.entur.netex.tools.lib.app and delete both copies. Update the four
consumers (cli Main, cli BasicTest, pipeline MainPipline, pipeline
TimetablePipeline) to import from the lib package.
